Fixed min/max snapping to be based off min if set not forced to zero
Issue:
If Step was set to 4, you could not set min to 1 to get a range like:
 [1 , 5 , 9 , 13 , 17 ...]

Fix was to set step to move off of min instead of zero as a fixed base.

Small bug but in interest of time was left for now. Min can not be incremented downwards by the down arrow, must be manually typed. (can move up and Max can move up or down)